BUG: Compute slice plane edge when computing PLANE_EDGE flag.
The slice rotation interaction relies on the PLANE_EDGE flag, which
ToolCursor sets when the mouse is within a certain distance of a slice
plane edge.  The old code was just computing the distance to the volume
boundary, which was insufficient: it is possible to be near the boundary
of the volume but nowhere near the edge of the slice plane.